+++ sistemski_softver/common/structures.h
@@ -86,7 +86,7 @@ public:
 	SymbolTableEntry* GetEntryByName(string name);
 
 	stringstream GenerateTextualSymbolTable();
-	unsigned long GetSize() { return table.size(); }
+	size_t GetSize() { return table.size(); }
 };
 
 /////////////////////////////////////////////////////////
@@ -122,7 +122,7 @@ public:
 	stringstream GenerateTextualSectionTable();
 
 	static string DefaultFlags(SectionType type);
-	unsigned long GetSize() { return table.size(); }
+	size_t GetSize() { return table.size(); }
 };
 
 ////////////////////////////////////////////////////////////
@@ -158,7 +158,7 @@ public:
 	
 	RelocationTableEntry* GetEntryByID(int id) { return &table.at(id); }
 	stringstream GenerateTextualRelocationTable();
-	unsigned long GetSize() { return table.size(); }
+	size_t GetSize() { return table.size(); }
 };
 
 #endif
